Fix for 64bit CPUs - use 32bit 'int' instead of 64bit 'long'

+++ ref-impl/aaflib/aafrdli.h
@@ -41,7 +41,7 @@
 typedef void * AAFLibraryHandle;
 typedef void * AAFSymbolAddr;
 typedef void * AAFUserData;
-typedef long AAFRDLIRESULT;
+typedef int AAFRDLIRESULT;
